clarify policy on committee conflict of interest
We have a known instance of committee conflict of interest with Code of
Conduct Committee members being elected to the Steering Committee.
Based on these bodies charter, membership is mutually exclusive.  But
it has not been clearly documented that in such cases one must choose
vacate their membership in one body or the other, or how the vacancy
would be filled.

This clarifying text documents our established practice.

Signed-off-by: Tim Pepper <tpepper@vmware.com>